REFACTOR: fixed routes middleware
