+++ src/routes/recipes.route.ts
@@ -2,7 +2,7 @@ import { Router } from 'express'
 import upload from '../middleware/multer'
 import recipeController from '../controllers/recipe.controller'
 import requireUser from '../middleware/require-user'
-import validateInput from '../middleware/validate-input'
+import validateInput from '../middleware/validateInput'
 import { createRecipeSchema, updateRecipeSchema } from '../schema/recipe.schema'
 
 const router = Router()
